+++ src/generate_value.cu
@@ -22,7 +22,7 @@ __global__ void assign<float>(float *ptr, int size, int var_idx)
     }
 }
 
-cudaError_t cuda_assign_double(MPI_Comm gcomm, float *ptr, int size, int var_idx)
+cudaError_t cuda_assign_double(MPI_Comm gcomm, double *ptr, int size, int var_idx)
 {
     MPI_Comm_rank(gcomm, &rank);
     int dev_num, dev_rank;
